requirements.txt
django
djangorestframework
sphinx